Fixed snake bug and updated rategaurd
Snake could move the opposite direction which would result in game over due to collision.
Resolved this with if statements in the handleKeydownEvent method. The RateGuardJS class had a 5 minute timeout in
script.js. This was too long. I reduced this to 2 minutes. Also fixed a comparison in the class object to make max requests
work correctly.